introducing new model architecture to try out
