+++ llmini/model.py
@@ -1,99 +1,46 @@
 # model.py
-import math
-import torch
-import torch.nn as nn
+from llmini.arch import TinyGPT, ComplexGPT
 
 
-class CausalSelfAttention(nn.Module):
-    def __init__(self, n_embd, n_head, block_size):
-        super().__init__()
-        assert n_embd % n_head == 0
-        self.n_head = n_head
-        self.key = nn.Linear(n_embd, n_embd, bias=False)
-        self.query = nn.Linear(n_embd, n_embd, bias=False)
-        self.value = nn.Linear(n_embd, n_embd, bias=False)
-        self.proj = nn.Linear(n_embd, n_embd)
-        self.register_buffer("mask", torch.tril(torch.ones(
-            block_size, block_size)).unsqueeze(0).unsqueeze(0))
-
-    def forward(self, x):
-        B, T, C = x.size()
-        k = self.key(x).view(B, T, self.n_head, C // self.n_head).transpose(1, 2)   # (B, nh, T, hs)
-        q = self.query(x).view(B, T, self.n_head, C // self.n_head).transpose(1, 2)
-        v = self.value(x).view(B, T, self.n_head, C // self.n_head).transpose(1, 2)
-        att = (q @ k.transpose(-2, -1)) / math.sqrt(k.size(-1))
-        att = att.masked_fill(self.mask[:, :, :T, :T] == 0, float('-inf'))
-        att = torch.softmax(att, dim=-1)
-        y = att @ v  # (B, nh, T, hs)
-        y = y.transpose(1, 2).contiguous().view(B, T, C)
-        return self.proj(y)
+def get_model(architecture, vocab_size, block_size, device):
+    if architecture == "tiny":
+        return TinyGPT(vocab_size, block_size=block_size, n_layer=6, n_head=8, n_embd=256, dropout=0.0).to(device)
+    elif architecture == "complex":
+        return ComplexGPT(vocab_size, block_size=block_size, n_layer=8, n_head=8, n_embd=512, dropout=0.1).to(device)
+    else:
+        raise ValueError(f"Unknown architecture: {architecture}")
 
 
-class Block(nn.Module):
-    def __init__(self, n_embd, n_head, block_size, dropout):
-        super().__init__()
-        self.ln1 = nn.LayerNorm(n_embd)
-        self.attn = CausalSelfAttention(n_embd, n_head, block_size)
-        self.ln2 = nn.LayerNorm(n_embd)
-        self.mlp = nn.Sequential(
-            nn.Linear(n_embd, 4 * n_embd),
-            nn.GELU(),
-            nn.Linear(4 * n_embd, n_embd),
-            nn.Dropout(dropout),
-        )
+def get_model_from_args(args, vocab_size, block_size, device):
+    """
+    Select and return the appropriate model based on the parsed arguments.
 
-    def forward(self, x):
-        x = x + self.attn(self.ln1(x))
-        x = x + self.mlp(self.ln2(x))
-        return x
+    Args:
+        args: Parsed command-line arguments.
+        vocab_size: Size of the vocabulary.
+        block_size: Maximum sequence length.
+        device: Device to load the model on.
 
-
-class TinyGPT(nn.Module):
-    def __init__(self, vocab_size, block_size=128, n_layer=4, n_head=4, n_embd=128, dropout=0.1):
-        super().__init__()
-        self.block_size = block_size
-        self.tok_emb = nn.Embedding(vocab_size, n_embd)
-        self.pos_emb = nn.Embedding(block_size, n_embd)
-        self.drop = nn.Dropout(dropout)
-        self.blocks = nn.Sequential(
-            *[Block(n_embd, n_head, block_size, dropout) for _ in range(n_layer)])
-        self.ln_f = nn.LayerNorm(n_embd)
-        self.head = nn.Linear(n_embd, vocab_size, bias=False)
-
-        self.apply(self._init_weights)
-
-    def _init_weights(self, m):
-        if isinstance(m, (nn.Linear, nn.Embedding)):
-            nn.init.xavier_uniform_(m.weight)  # Use Xavier initialization
-        if isinstance(m, nn.Linear) and m.bias is not None:
-            nn.init.zeros_(m.bias)
-
-    def forward(self, idx, targets=None):
-        B, T = idx.size()
-        pos = torch.arange(0, T, device=idx.device).unsqueeze(0)
-        x = self.tok_emb(idx) + self.pos_emb(pos)
-        x = self.drop(x)
-        x = self.blocks(x)
-        x = self.ln_f(x)
-        logits = self.head(x)
-        loss = None
-        if targets is not None:
-            loss = torch.nn.functional.cross_entropy(
-                logits.view(-1, logits.size(-1)), targets.view(-1))
-        return logits, loss
-
-    @torch.no_grad()
-    def generate(self, idx, max_new_tokens, temperature=1.0, top_k=None):
-        for _ in range(max_new_tokens):
-            idx_cond = idx[:, -self.block_size:]
-            logits, _ = self(idx_cond)
-            logits = logits[:, -1, :] / temperature
-            if top_k is not None:
-                k = min(top_k, logits.size(-1))
-                vals, idxs = torch.topk(logits, k)
-                mask = torch.full_like(logits, float('-inf'))
-                logits = mask.scatter(1, idxs, vals)
-            probs = torch.softmax(logits, dim=-1)
-            next_id = torch.multinomial(probs, num_samples=1)
-            idx = torch.cat((idx, next_id), dim=1)
-        return idx
+    Returns:
+        An instance of the selected model.
+    """
+    if args.model == "tiny":
+        return TinyGPT(
+            vocab_size=vocab_size,
+            block_size=block_size,
+            n_layer=6,
+            n_head=8,
+            n_embd=256,
+            dropout=0.0
+        ).to(device)
+    elif args.model == "complex":
+        return ComplexGPT(
+            vocab_size=vocab_size,
+            block_size=block_size,
+            n_layer=8,
+            n_head=8,
+            n_embd=512,
+            dropout=0.1
+        ).to(device)
+    else:
+        raise ValueError(f"Unknown model type: {args.model}")
