+++ llmini/utils.py
@@ -1,3 +1,4 @@
+import argparse
 import logging
 
 
@@ -10,3 +11,10 @@ def setup_logger(name, level=logging.INFO):
     handler.setFormatter(formatter)
     logger.addHandler(handler)
     return logger
+
+
+def parse_arguments():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--model", choices=["tiny", "complex"], default="tiny",
+                        help="Choose the model architecture: 'tiny' or 'complex'")
+    return parser.parse_args()
